Backend-venv-flask/requirements.txt
flask
flask-cors
gunicorn==19.6.0
google-cloud